requirements.txt
telethon==1.40.0
streamlit==1.37.0
SQLModel==0.0.21
loguru==0.7.2
python-dotenv==1.0.1
rich==13.7.1
pydantic==2.8.2
uvloop==0.19.0; platform_system != "Windows"
pandas==2.2.2
pyarrow==16.1.0
psycopg2-binary==2.9.9
